refactor: move fortran inference helpers
